pyproject.toml
[tool.poetry]
name = "healthyhabitsdrf(spa)"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
pillow = "^11.1.0"
djangorestframework-simplejwt = "^5.4.0"
requests = "^2.32.3"
redis = "^5.2.1"
celery = "^5.4.0"
eventlet = "^0.39.0"
django-celery-beat = "^2.7.0"
pytz = "^2025.1"
django-cors-headers = "^4.7.0"
drf-yasg = "^1.21.8"
coverage = "^7.6.12"
black = "^25.1.0"
isort = "^6.0.0"
flake8 = "^7.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
exclude = '''
(
  /(
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
)
'''

[tool.isort]
line_length = 119
